I added background music from a asset pack in unity called A Hollow Darkness.
Made all the Audio varibles public .

Assigned them into unity.

Renamed the variables to a more appropiate name.

I called the audio sorce from the ball and got its componets.

I then added the code to allow the audio clips to pplay depending on if it hit the wall, player, or was scored.